commit 01/08/2024: de agregan las siguientes entidades: empresa, grupo, horario-laboratorio, laboratorio, paciente, paquete, unidad. Además de ajustes al servicio de usuario para que este lo mas parecido al esquema q paso Zaira

+++ src/main/java/com/api/wslaboratorio/entities/CitaEntity.java
@@ -36,7 +36,7 @@ public class CitaEntity {
             )
     @Column(name = "citaid")
     @Comment("ID AUTOGENERADO PARA LOGRAR LA ATOMICIDAD DEL REGISTRO.")
-    private Integer citaId;
+    private Long citaId;
 
     @Embedded
     private AuditoriaEntity auditoriaEntity;
